move SQLiteSerialization into its own file for clarity
